more node creation tweaks for speed
git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/trunk@745 a380766d-ff14-0410-b294-a243070f3f08

+++ headers/plugin_interface/SC_Wire.h
@@ -27,7 +27,7 @@
 struct Wire 
 {
 	struct Unit *mFromUnit;
-	int16 mCalcRate;
+	int32 mCalcRate;
 	float32 *mBuffer;
 	float32 mScalarValue;
 };
